feat: CheerButton 컴포넌트의 클릭 핸들링 로직을 useCheerButton 훅으로 분리
