+++ frontend/src/domains/components/CheerButton/CheerButton.tsx
@@ -1,4 +1,3 @@
-import { postOrganizationCheer } from '@/apis/organization.api';
 import Button from '@/components/@commons/Button/Button';
 import FillHeartIcon from '@/components/icons/FillHeartIcon';
 import {
@@ -7,38 +6,15 @@ import {
   iconWrapperStyle,
   textStyle,
 } from '@/domains/components/CheerButton/CheerButton.style';
-import { useDebounce } from '@/hooks/useDebounce';
+import useCheerButton from '@/domains/hooks/useCheerButton';
 import { theme } from '@/theme';
-import { useState } from 'react';
 
 interface CheerButtonProps {
   totalCheeringCount: number;
 }
 
 export default function CheerButton({ totalCheeringCount }: CheerButtonProps) {
-  const [count, setCount] = useState(0);
-  const [accCount, setAccCount] = useState(0);
-  const [animate, setAnimate] = useState(false);
-
-  const handleClick = () => {
-    setCount(count + 1);
-    setAccCount(accCount + 1);
-    setAnimate(false);
-    requestAnimationFrame(() => setAnimate(true));
-
-    debouncedSearch(count + 1);
-  };
-
-  const debouncedSearch = useDebounce(
-    ((count: number) => {
-      postOrganizationCheer({
-        organizationId: 1,
-        cheeringCount: count,
-      });
-      setCount(0);
-    }) as (...args: unknown[]) => void,
-    500
-  );
+  const { handleClick, animate, accCount } = useCheerButton();
 
   return (
     <Button onClick={handleClick} css={cheerButtonStyle}>
